allow easy extension of startup scene
